+++ frontend/src/modules/parseTaskText/index.ts
@@ -1,0 +1,5 @@
+export {parseTaskText} from './parseTaskText'
+export {PrefixMode, PREFIXES} from './prefixes'
+export {getLabelsFromPrefix, getProjectFromPrefix} from './prefixParser'
+export {cleanupItemText} from './textCleanup'
+export type {ParsedTaskText} from './types'
